feat(prompt): add prompt
add prompt
